ActivitySceneTransitionBasic: AndroidX
Migrate to AndroidX and update dependencies.
