--- ActivitySceneTransitionBasic/Application/build.gradle
@@ -1,69 +1,0 @@
-
-buildscript {
-    repositories {
-        google()
-        jcenter()
-    }
-
-    dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.2'
-    }
-}
-
-apply plugin: 'com.android.application'
-
-repositories {
-    google()
-    jcenter()
-}
-
-dependencies {
-
-
-        implementation "com.android.support:support-v4:28.0.0"
-        implementation "com.android.support:support-v13:28.0.0"
-        implementation "com.android.support:cardview-v7:28.0.0"
-        implementation "com.android.support:appcompat-v7:28.0.0"
-
-
-    implementation 'com.squareup.picasso:picasso:2.4.0'
-
-
-
-
-}
-
-// The sample build uses multiple directories to
-// keep boilerplate and common code separate from
-// the main sample code.
-List<String> dirs = [
-    'main',     // main sample code; look here for the interesting stuff.
-    'common',   // components that are reused by multiple samples
-    'template'] // boilerplate code that is generated by the sample template process
-
-android {
-    compileSdkVersion 28
-
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 28
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
-    }
-
-    sourceSets {
-        main {
-            dirs.each { dir ->
-                java.srcDirs "src/${dir}/java"
-                res.srcDirs "src/${dir}/res"
-            }
-        }
-        androidTest.setRoot('tests')
-        androidTest.java.srcDirs = ['tests/src']
-
-    }
-
-}
