--- ActivitySceneTransitionBasic/Application/src/main/res/layout/grid_item.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright (C) 2014 The Android Open Source Project
-
-     Licensed under the Apache License, Version 2.0 (the "License");
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content"
-              android:orientation="vertical">
-
-    <com.example.android.activityscenetransitionbasic.SquareFrameLayout
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content">
-
-        <ImageView
-              android:id="@+id/imageview_item"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent"
-              android:scaleType="centerCrop" />
-
-    </com.example.android.activityscenetransitionbasic.SquareFrameLayout>
-
-    <TextView
-          android:id="@+id/textview_name"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:background="?android:attr/colorPrimary"
-          android:theme="@android:style/Theme.Material"
-          android:textAppearance="@android:style/TextAppearance.Material.Subhead"
-          android:maxLines="1"
-          android:padding="16dp" />
-
-</LinearLayout>
