--- ActivitySceneTransitionBasic/Application/src/main/res/values-v21/styles.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright (C) 2014 The Android Open Source Project
-
-     Licensed under the Apache License, Version 2.0 (the "License");
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
-
-<resources>
-
-    <style name="Theme.Sample" parent="Theme.Base">
-        <!-- Set the transitions which are used for the entrance and exit of shared elements -->
-        <item name="android:windowSharedElementEnterTransition">
-            @transition/grid_detail_transition
-        </item>
-        <item name="android:windowSharedElementExitTransition">
-            @transition/grid_detail_transition
-        </item>
-    </style>
-
-</resources>
