--- ActivitySceneTransitionBasic/Application/src/main/res/values/base-strings.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
- Copyright 2013 The Android Open Source Project
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--->
-
-<resources>
-    <string name="app_name">ActivitySceneTransitionBasic</string>
-    <string name="intro_message">
-        <![CDATA[
-        
-            
-            Demonstrates how to the use Activity scene transitions when transitions
-            from one Activity to another. Uses a combination of changeImageTransform and
-            changeBounds to nicely transition a grid of images to an Activity with a large image
-            and detail text.
-            
-        
-        ]]>
-    </string>
-</resources>
