--- ActivitySceneTransitionBasic/Application/src/main/res/values/strings.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright (C) 2014 The Android Open Source Project
-
-     Licensed under the Apache License, Version 2.0 (the "License");
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
-
-<resources>
-
-    <string name="bacon_ipsum">
-        Bacon ipsum dolor sit amet venison shankle pork chop meatball tri-tip beef ribs turducken.
-        Strip steak ball tip boudin shank, turducken leberkas pork chop beef ribs ham hock sausage
-        frankfurter prosciutto doner ham. Bacon landjaeger ball tip, andouille chuck beef ribs jowl
-        kevin tri-tip turkey biltong frankfurter sausage. Boudin pork belly meatloaf chicken cow
-        tri-tip kielbasa shoulder. Pork loin pig boudin hamburger pastrami short ribs. Sirloin
-        tongue pork loin chicken spare ribs bresaola pastrami.\n\n
-        Tenderloin turducken pork chop, pork belly beef ribs brisket ham. Turducken landjaeger short
-        loin capicola pancetta pork chop strip steak rump meatloaf brisket kevin doner short ribs
-        salami. Beef prosciutto flank leberkas landjaeger swine. Fatback prosciutto sausage, jerky
-        tail tongue hamburger jowl biltong shank pork belly swine filet mignon chicken ground round.
-        Pork chop porchetta ground round tri-tip tail t-bone.\n\n
-        Landjaeger rump bacon salami sausage pork loin pig brisket strip steak corned beef. Biltong
-        sirloin meatloaf ribeye, bresaola cow chicken t-bone frankfurter andouille strip steak jerky
-        capicola. Ribeye porchetta strip steak boudin. Kielbasa cow brisket pastrami ball tip
-        tenderloin bresaola ham. Biltong andouille chuck ham hock jerky beef chicken, flank shankle
-        ball tip venison porchetta kevin fatback kielbasa. Boudin tongue ground round, turkey pork
-        belly salami corned beef pork tri-tip meatloaf sausage andouille strip steak pig. Spare ribs
-        beef meatloaf rump sausage doner frankfurter.\n\n
-        Fatback porchetta pork loin meatball, turducken pork chop drumstick boudin. Kevin tri-tip
-        ground round corned beef, ribeye swine filet mignon salami pork spare ribs pork chop
-        brisket. Filet mignon shankle t-bone bacon. Turducken capicola turkey porchetta kielbasa
-        shank pork loin jerky venison tenderloin boudin ham hock ground round.</string>
-
-    <string name="image_header">%1$s by %2$s</string>
-
-</resources>
