+++ ActivitySceneTransitionBasic/app/src/main/res/drawable-xxhdpi/ic_launcher.png
