+++ ActivitySceneTransitionBasic/app/src/main/res/transition/grid_detail_transition.xml
