+++ ActivitySceneTransitionBasic/app/src/main/res/values/colors.xml
