+++ ActivitySceneTransitionBasic/gradle.properties
@@ -18,3 +18,5 @@
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
 
+android.enableJetifier=true
+android.useAndroidX=true
