+++ ActivitySceneTransitionBasic/gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,6 @@
+#Tue Aug 27 15:13:39 JST 2019
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.1.1-all.zip
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.4.1-all.zip
